+++ Sources/MapboxCoreNavigation/LegacyRouteController.swift
@@ -51,6 +51,13 @@ open class LegacyRouteController: NSObject, Router, InternalRouter, CLLocationMa
 
     // MARK: Tracking the Progress
     
+    /**
+     Provides notifications and access to `AlternativeRoute`s found during navigation.
+     
+     - important: Always `nil` for `LegacyRouteController`.
+     */
+    public private(set) var alternativesStore: NavigatorAlternativesStore? = nil
+    
     public weak var delegate: RouterDelegate?
     
     public internal(set) var routeProgress: RouteProgress {